npc_be/src/requirements.txt

fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
passlib[bcrypt]
bcrypt
python-jose[cryptography]
psycopg2-binary
sqlalchemy
pyotp
